Title: 특정 숫자 도달하기, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240729/특정 숫자 도달하기/reaching-specific-number.c
@@ -12,7 +12,7 @@ int main() {
         count += 1;
     }
 
-    double avg = sum / count;
+    double avg = (double)sum / count;
 
     printf("%d %.1lf", sum, avg);
